refactor: restructure e2e tests with page.route() mocking and consolidate test folders
- Migrate all e2e tests from mock server to Playwright route interception
- Remove mock server dependency from e2e (webServer config, mock:server:e2e script)
- Reduce Playwright projects from 6 to 3 (one per engine: Chromium, Firefox, WebKit)
- Split mega main-flow test into focused tests with API-seeded data
- Merge scroll-restore test into Item CRUD describe block
- Move all tests to tests/ folder (tests/unit, tests/e2e, tests/integration)
- Fix scroll restore retry logic for WebKit CI compatibility
- Update test scripts: npm test runs unit + e2e, npm run test:unit for unit only
