+++ .husky/pre-commit
@@ -9,6 +9,6 @@ npm run typecheck
 echo "Running lint-staged..."
 npx lint-staged
 
-# Run tests
-echo "Running tests..."
-npm run test:run
+# Run unit tests
+echo "Running unit tests..."
+npm run test:unit
