+++ package.json
@@ -7,15 +7,15 @@
     "predev": "npm run api:fetch",
     "dev": "vite",
     "build": "vite build",
-    "test": "vitest",
-    "test:run": "vitest run",
+    "test": "vitest run && playwright test",
+    "test:unit": "vitest run",
+    "test:unit:watch": "vitest",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "typecheck": "tsc --noEmit",
     "prepare": "husky",
     "mock:server": "tsx watch api/server.ts",
     "mock:server:ci": "tsx api/server.ts",
-    "mock:server:e2e": "MOCK_PERSIST=false tsx api/server.ts",
     "routes": "npx @tanstack/router-cli@latest generate",
     "api:fetch": "curl -s -o src/core/openapi.json https://raw.githubusercontent.com/Alexgub84/chillist-be/main/docs/openapi.json",
     "api:types": "openapi-typescript src/core/openapi.json -o src/core/api.generated.ts && prettier --write src/core/api.generated.ts",
