+++ playwright.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig, devices } from '@playwright/test';
 
 export default defineConfig({
-  testDir: './e2e',
+  testDir: './tests/e2e',
   fullyParallel: true,
   forbidOnly: !!process.env.CI,
   retries: process.env.CI ? 2 : 0,
@@ -22,34 +22,14 @@ export default defineConfig({
       use: { ...devices['Desktop Firefox'] },
     },
     {
-      name: 'iPhone 15 Pro',
+      name: 'Mobile Safari',
       use: { ...devices['iPhone 15 Pro'] },
     },
-    {
-      name: 'iPhone SE',
-      use: { ...devices['iPhone SE'] },
-    },
-    {
-      name: 'Pixel 7',
-      use: { ...devices['Pixel 7'] },
-    },
-    {
-      name: 'Galaxy S9+',
-      use: { ...devices['Galaxy S9+'] },
-    },
   ],
-  webServer: [
-    {
-      command: 'npm run mock:server:e2e',
-      url: 'http://localhost:3333/health',
-      reuseExistingServer: !process.env.CI,
-      timeout: 30000,
-    },
-    {
-      command: 'npx vite',
-      url: 'http://localhost:5173',
-      reuseExistingServer: !process.env.CI,
-      timeout: 120000,
-    },
-  ],
+  webServer: {
+    command: 'npx vite',
+    url: 'http://localhost:5173',
+    reuseExistingServer: !process.env.CI,
+    timeout: 120000,
+  },
 });
