+++ tests/unit/api/server.test.ts
@@ -1,0 +1,247 @@
+import { describe, expect, it } from 'vitest';
+
+import { buildServer } from '../../../api/server';
+import type { MockData } from '../../../api/mock';
+
+function createTestData(): MockData {
+  const now = new Date().toISOString();
+  return {
+    plans: [
+      {
+        planId: 'plan-1',
+        title: 'Test Plan',
+        status: 'draft',
+        visibility: 'private',
+        ownerParticipantId: 'participant-1',
+        participantIds: ['participant-1'],
+        createdAt: now,
+        updatedAt: now,
+      },
+    ],
+    participants: [
+      {
+        participantId: 'participant-1',
+        displayName: 'Alex',
+        name: 'Alex',
+        lastName: 'Guberman',
+        role: 'owner',
+        isOwner: true,
+        createdAt: now,
+        updatedAt: now,
+      },
+    ],
+    items: [
+      {
+        itemId: 'item-1',
+        planId: 'plan-1',
+        name: 'Tent',
+        quantity: 1,
+        unit: 'pcs',
+        status: 'pending',
+        category: 'equipment',
+        createdAt: now,
+        updatedAt: now,
+      },
+    ],
+  };
+}
+
+describe('mock server', () => {
+  it('lists existing plans', async () => {
+    const server = await buildServer({
+      initialData: createTestData(),
+      persist: false,
+      logger: false,
+    });
+    try {
+      const response = await server.inject({ method: 'GET', url: '/plans' });
+      expect(response.statusCode).toBe(200);
+
+      const payload = response.json() as unknown;
+      expect(Array.isArray(payload)).toBe(true);
+      expect((payload as Array<Record<string, unknown>>).length).toBe(1);
+    } finally {
+      await server.close();
+    }
+  });
+
+  it('creates a new plan with defaults', async () => {
+    const server = await buildServer({
+      initialData: createTestData(),
+      persist: false,
+      logger: false,
+    });
+    try {
+      const response = await server.inject({
+        method: 'POST',
+        url: '/plans',
+        payload: {
+          title: 'Summer Trip',
+        },
+      });
+
+      expect(response.statusCode).toBe(201);
+      const payload = response.json() as Record<string, unknown>;
+      expect(payload.planId).toEqual(expect.any(String));
+      expect(payload.status).toBe('draft');
+      expect(payload.visibility).toBe('private');
+    } finally {
+      await server.close();
+    }
+  });
+
+  it('fetches a single plan with items', async () => {
+    const server = await buildServer({
+      initialData: createTestData(),
+      persist: false,
+      logger: false,
+    });
+    try {
+      const response = await server.inject({
+        method: 'GET',
+        url: '/plans/plan-1',
+      });
+      expect(response.statusCode).toBe(200);
+
+      const payload = response.json() as Record<string, unknown>;
+      expect(payload.planId).toBe('plan-1');
+      expect(payload.title).toBe('Test Plan');
+      expect(Array.isArray(payload.items)).toBe(true);
+      expect((payload.items as Array<Record<string, unknown>>).length).toBe(1);
+      expect((payload.items as Array<Record<string, unknown>>)[0].itemId).toBe(
+        'item-1'
+      );
+    } finally {
+      await server.close();
+    }
+  });
+
+  it('adds a participant to a plan', async () => {
+    const server = await buildServer({
+      initialData: createTestData(),
+      persist: false,
+      logger: false,
+    });
+    try {
+      const response = await server.inject({
+        method: 'POST',
+        url: '/plans/plan-1/participants',
+        payload: {
+          name: 'Jamie',
+          lastName: 'Rivera',
+          contactPhone: '+1234567890',
+        },
+      });
+
+      expect(response.statusCode).toBe(201);
+      const payload = response.json() as Record<string, unknown>;
+      expect(payload.participantId).toEqual(expect.any(String));
+      expect(payload.role).toBe('participant');
+
+      const planResponse = await server.inject({
+        method: 'GET',
+        url: '/plans/plan-1',
+      });
+      const plan = planResponse.json() as { participantIds?: string[] };
+      expect(plan.participantIds).toEqual(
+        expect.arrayContaining([payload.participantId as string])
+      );
+    } finally {
+      await server.close();
+    }
+  });
+
+  it('creates a plan with owner via /plans/with-owner', async () => {
+    const server = await buildServer({
+      initialData: createTestData(),
+      persist: false,
+      logger: false,
+    });
+    try {
+      const response = await server.inject({
+        method: 'POST',
+        url: '/plans/with-owner',
+        payload: {
+          title: 'Beach Trip',
+          owner: {
+            name: 'Alex',
+            lastName: 'G',
+            contactPhone: '+9876543210',
+          },
+          participants: [
+            {
+              name: 'Jamie',
+              lastName: 'Rivera',
+              contactPhone: '+1111111111',
+            },
+          ],
+        },
+      });
+
+      expect(response.statusCode).toBe(201);
+      const payload = response.json() as Record<string, unknown>;
+      expect(payload.planId).toEqual(expect.any(String));
+      expect(payload.title).toBe('Beach Trip');
+      expect(payload.status).toBe('draft');
+      expect(Array.isArray(payload.items)).toBe(true);
+      expect((payload.items as unknown[]).length).toBe(0);
+      expect(Array.isArray(payload.participants)).toBe(true);
+      const participants = payload.participants as Array<
+        Record<string, unknown>
+      >;
+      expect(participants.length).toBe(2);
+      expect(participants[0].role).toBe('owner');
+      expect(participants[0].name).toBe('Alex');
+      expect(participants[1].role).toBe('participant');
+      expect(participants[1].name).toBe('Jamie');
+    } finally {
+      await server.close();
+    }
+  });
+
+  it('creates a new item for a plan', async () => {
+    const server = await buildServer({
+      initialData: createTestData(),
+      persist: false,
+      logger: false,
+    });
+    try {
+      const response = await server.inject({
+        method: 'POST',
+        url: '/plans/plan-1/items',
+        payload: {
+          name: 'Lantern',
+          category: 'equipment',
+          quantity: 2,
+          status: 'pending',
+        },
+      });
+
+      expect(response.statusCode).toBe(201);
+      const payload = response.json() as Record<string, unknown>;
+      expect(payload.itemId).toEqual(expect.any(String));
+      expect(payload.planId).toBe('plan-1');
+      expect(payload.quantity).toBe(2);
+      expect(payload.status).toBe('pending');
+    } finally {
+      await server.close();
+    }
+  });
+
+  it('returns 404 when a plan is missing', async () => {
+    const server = await buildServer({
+      initialData: createTestData(),
+      persist: false,
+      logger: false,
+    });
+    try {
+      const response = await server.inject({
+        method: 'GET',
+        url: '/plans/unknown',
+      });
+      expect(response.statusCode).toBe(404);
+    } finally {
+      await server.close();
+    }
+  });
+});
