+++ vite.config.ts
@@ -8,7 +8,7 @@ export default defineConfig({
   test: {
     globals: true,
     environment: 'jsdom',
-    setupFiles: './src/test/setup.ts',
-    exclude: ['**/node_modules/**', '**/e2e/**'],
+    setupFiles: './tests/setup.ts',
+    exclude: ['**/node_modules/**', '**/tests/e2e/**'],
   },
 });
